requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0

# Gmail API
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# Slack API
slack-sdk==3.26.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# CORS
python-jose[cryptography]==3.3.0

